Revert "Refactor/wallets 2"
