+++ src/components/Forms/AddWalletFormFields.tsx
@@ -1,6 +1,7 @@
 import { type QRL, component$, $ } from "@builder.io/qwik";
 import { getAddress } from "viem";
 import { Input } from "~/components/Input/Input";
+import { type addWalletFormStore } from "~/routes/app/wallets";
 import { useDebouncer } from "~/utils/debouncer";
 import {
   isCheckSum,
@@ -11,10 +12,9 @@ import {
 import IconSuccess from "/public/assets/icons/dashboard/success.svg?jsx";
 import IconWarning from "/public/assets/icons/dashboard/warning.svg?jsx";
 import { Button } from "../Buttons/Buttons";
-import { type AddWalletFormStore } from "~/interface/wallets/addWalletFormStore";
 
 export interface AddWalletFormFieldsProps {
-  addWalletFormStore: AddWalletFormStore;
+  addWalletFormStore: addWalletFormStore;
   onConnectWalletClick: QRL<() => void>;
   isWalletConnected: boolean | undefined;
 }
