+++ src/components/Forms/AmountOfCoins.tsx
@@ -1,12 +1,15 @@
 import { type Signal, component$ } from "@builder.io/qwik";
 import { FormBadge } from "~/components/FormBadge/FormBadge";
+
+import {
+  chekckIfProperAmount,
+  replaceNonMatching,
+  type addWalletFormStore,
+} from "~/routes/app/wallets";
 import { Button } from "../Buttons/Buttons";
-import { replaceNonMatching } from "~/utils/replaceNonMatching";
-import { checkIfStringMatchesPattern } from "~/utils/checkIfStringMatchesPattern";
-import { type AddWalletFormStore } from "~/interface/wallets/addWalletFormStore";
 
 export interface AmountOfCoinsProps {
-  addWalletFormStore: AddWalletFormStore;
+  addWalletFormStore: addWalletFormStore;
   walletTokenBalances: Signal<any>;
 }
 
@@ -80,7 +83,7 @@ export default component$<AmountOfCoinsProps>(
                 />
               </div>
               <span class="block pb-1 text-xs text-white">
-                {!checkIfStringMatchesPattern(
+                {!chekckIfProperAmount(
                   addWalletFormStore.coinsToApprove.find(
                     (item) => item.symbol === symbol,
                   )!.amount,
