+++ src/components/Forms/CoinsToApprove.tsx
@@ -1,9 +1,10 @@
 import { type Signal, component$ } from "@builder.io/qwik";
 import { FormBadge } from "~/components/FormBadge/FormBadge";
-import { type AddWalletFormStore } from "~/interface/wallets/addWalletFormStore";
+
+import { type addWalletFormStore } from "~/routes/app/wallets";
 
 export interface CoinsToApproveProps {
-  addWalletFormStore: AddWalletFormStore;
+  addWalletFormStore: addWalletFormStore;
   walletTokenBalances: Signal<any>;
 }
