+++ src/components/Forms/portfolioTransfters/CoinsAmounts.tsx
@@ -1,9 +1,8 @@
 import { component$ } from "@builder.io/qwik";
 import { FormBadge } from "~/components/FormBadge/FormBadge";
 import { type BatchTransferFormStore } from "~/routes/app/portfolio";
+import { chekckIfProperAmount, replaceNonMatching } from "~/routes/app/wallets";
 import IconArrowDown from "/public/assets/icons/arrow-down.svg?jsx";
-import { checkIfStringMatchesPattern } from "~/utils/checkIfStringMatchesPattern";
-import { replaceNonMatching } from "~/utils/replaceNonMatching";
 
 export interface CoinsAmountsProps {
   batchTransferFormStore: BatchTransferFormStore;
@@ -42,10 +41,7 @@ export default component$<CoinsAmountsProps>(({ batchTransferFormStore }) => {
                           image={`/assets/icons/tokens/${coin.symbol.toLowerCase()}.svg`}
                           // for={}
                           hasImg={
-                            checkIfStringMatchesPattern(
-                              coin.amount,
-                              /^\d*\.?\d*$/,
-                            ) &&
+                            chekckIfProperAmount(coin.amount, /^\d*\.?\d*$/) &&
                             coin.amount != "0" &&
                             coin.amount.length > 0 &&
                             coin.amount[0] != "0"
@@ -57,7 +53,7 @@ export default component$<CoinsAmountsProps>(({ batchTransferFormStore }) => {
                               type="text"
                               name={`${coin.symbol}Amount`}
                               class={`custom-border-1 absolute end-4 block h-8 w-1/2 rounded-lg bg-transparent p-3 text-sm  placeholder-white placeholder-opacity-50 
-                              ${checkIfStringMatchesPattern(coin.amount, /^\d*\.?\d*$/) && coin.amount != "0" && coin.amount.length > 0 && coin.amount[0] != "0" ? "border-[#24a148] text-[#24a148] focus:border-[#24a148]" : ""}`}
+                              ${chekckIfProperAmount(coin.amount, /^\d*\.?\d*$/) && coin.amount != "0" && coin.amount.length > 0 && coin.amount[0] != "0" ? "border-[#24a148] text-[#24a148] focus:border-[#24a148]" : ""}`}
                               placeholder={`${coin.symbol} approval limit`}
                               value={coin.amount}
                               onInput$={(e) => {
@@ -87,10 +83,7 @@ export default component$<CoinsAmountsProps>(({ batchTransferFormStore }) => {
                         <span class="block pb-1 text-xs text-white">
                           {coin.amount.length < 1 &&
                           coin.amount[0] != "0" &&
-                          !checkIfStringMatchesPattern(
-                            coin.amount,
-                            /^\d*\.?\d*$/,
-                          ) ? (
+                          !chekckIfProperAmount(coin.amount, /^\d*\.?\d*$/) ? (
                             <span class="text-xs text-red-500">
                               Invalid amount. There should be only one dot.
                             </span>
