+++ src/interface/streamStore/streamStore.ts
@@ -1,9 +1,9 @@
 import { createContextId } from "@builder.io/qwik";
 
-export interface StreamStore {
+export interface ModalStore {
   streamId: string;
 }
 
-export const StreamStoreContext = createContextId<StreamStore>(
+export const StreamStoreContext = createContextId<ModalStore>(
   "stream-store-context",
 );
