--- src/interface/token/TransferredCoin.ts
@@ -1,4 +1,0 @@
-export interface TransferredCoinInterface {
-  symbol: string;
-  address: string;
-}
